cmd/atlas/internal/cmdapi: allow extending schema inspect
