+++ cmd/atlas/internal/cmdapi/cmdapi.go
@@ -464,6 +464,7 @@ type stateReaderConfig struct {
 	client, dev *sqlclient.Client // database connections, while dev is considered a dev database, client is not
 	schemas     []string          // schemas to work on
 	exclude     []string          // exclude flag values
+	include     []string          // include flag values
 	withPos     bool              // indicate if schema.Pos should be loaded.
 	vars        Vars
 }
@@ -485,6 +486,7 @@ func (c *stateReaderConfig) Exported() (*cmdext.StateReaderConfig, error) {
 		Dev:     c.dev,
 		Schemas: c.schemas,
 		Exclude: c.exclude,
+		Include: c.include,
 		WithPos: c.withPos,
 		Vars:    c.vars,
 	}, nil
@@ -574,9 +576,13 @@ func stateReader(ctx context.Context, env *Env, config *stateReaderConfig) (*cmd
 			sr = migrate.RealmConn(c.Driver, &schema.InspectRealmOption{
 				Schemas: config.schemas,
 				Exclude: config.exclude,
+				Include: config.include,
 			})
 		default:
-			sr = migrate.SchemaConn(c.Driver, c.URL.Schema, &schema.InspectOptions{Exclude: config.exclude})
+			sr = migrate.SchemaConn(c.Driver, c.URL.Schema, &schema.InspectOptions{
+				Exclude: config.exclude,
+				Include: config.include,
+			})
 		}
 		return &cmdext.StateReadCloser{
 			StateReader: sr,
