+++ cmd/atlas/internal/cmdapi/schema.go
@@ -346,113 +346,6 @@ func schemaDiffRun(cmd *cobra.Command, _ []string, flags schemaDiffFlags, env *E
 	)
 }
 
-type schemaInspectFlags struct {
-	url       string   // URL of resource to inspect.
-	devURL    string   // URL of the dev database.
-	logFormat string   // Format of the log output.
-	schemas   []string // Schemas to take into account when diffing.
-	exclude   []string // List of glob patterns used to filter resources from applying (see schema.InspectOptions).
-}
-
-// schemaInspectCmd represents the 'atlas schema inspect' subcommand.
-func schemaInspectCmd() *cobra.Command {
-	cmd, _ := schemaInspectCmdWithFlags()
-	return cmd
-}
-
-func schemaInspectCmdWithFlags() (*cobra.Command, *schemaInspectFlags) {
-	var (
-		env   *Env
-		flags schemaInspectFlags
-		cmd   = &cobra.Command{
-			Use:   "inspect",
-			Short: "Inspect a database and print its schema in Atlas DDL syntax.",
-			Long: `'atlas schema inspect' connects to the given database and inspects its schema.
-It then prints to the screen the schema of that database in Atlas DDL syntax. This output can be
-saved to a file, commonly by redirecting the output to a file named with a ".hcl" suffix:
-
-  atlas schema inspect -u "mysql://user:pass@localhost:3306/dbname" > schema.hcl
-
-This file can then be edited and used with the` + " `atlas schema apply` " + `command to plan
-and execute schema migrations against the given database. In cases where users wish to inspect
-all multiple schemas in a given database (for instance a MySQL server may contain multiple named
-databases), omit the relevant part from the url, e.g. "mysql://user:pass@localhost:3306/".
-To select specific schemas from the databases, users may use the "--schema" (or "-s" shorthand)
-flag.
-	`,
-			Example: `  atlas schema inspect -u "mysql://user:pass@localhost:3306/dbname"
-  atlas schema inspect -u "mariadb://user:pass@localhost:3306/" --schema=schemaA,schemaB -s schemaC
-  atlas schema inspect --url "postgres://user:pass@host:port/dbname?sslmode=disable"
-  atlas schema inspect -u "sqlite://file:ex1.db?_fk=1"`,
-			PreRunE: RunE(func(cmd *cobra.Command, args []string) (err error) {
-				if env, err = selectEnv(cmd); err != nil {
-					return err
-				}
-				return setSchemaEnvFlags(cmd, env)
-			}),
-			RunE: RunE(func(cmd *cobra.Command, args []string) error {
-				return schemaInspectRun(cmd, args, flags, env)
-			}),
-		}
-	)
-	cmd.Flags().SortFlags = false
-	addFlagURL(cmd.Flags(), &flags.url)
-	addFlagDevURL(cmd.Flags(), &flags.devURL)
-	addFlagSchemas(cmd.Flags(), &flags.schemas)
-	addFlagExclude(cmd.Flags(), &flags.exclude)
-	addFlagLog(cmd.Flags(), &flags.logFormat)
-	addFlagFormat(cmd.Flags(), &flags.logFormat)
-	cobra.CheckErr(cmd.MarkFlagRequired(flagURL))
-	cmd.MarkFlagsMutuallyExclusive(flagLog, flagFormat)
-	return cmd, &flags
-}
-
-func schemaInspectRun(cmd *cobra.Command, _ []string, flags schemaInspectFlags, env *Env) error {
-	var (
-		ctx = cmd.Context()
-		dev *sqlclient.Client
-	)
-	useDev, err := readerUseDev(env, flags.url)
-	if err != nil {
-		return err
-	}
-	if flags.devURL != "" && useDev {
-		if dev, err = sqlclient.Open(ctx, flags.devURL); err != nil {
-			return err
-		}
-		defer dev.Close()
-	}
-	r, err := stateReader(ctx, env, &stateReaderConfig{
-		urls:    []string{flags.url},
-		dev:     dev,
-		vars:    env.Vars(),
-		schemas: flags.schemas,
-		exclude: flags.exclude,
-	})
-	if err != nil {
-		return err
-	}
-	defer r.Close()
-	client, ok := r.Closer.(*sqlclient.Client)
-	if !ok && dev != nil {
-		client = dev
-	}
-	format := cmdlog.SchemaInspectTemplate
-	if v := flags.logFormat; v != "" {
-		if format, err = template.New("format").Funcs(cmdlog.InspectTemplateFuncs).Parse(v); err != nil {
-			return fmt.Errorf("parse log format: %w", err)
-		}
-	}
-	s, err := r.ReadState(ctx)
-	if err != nil {
-		return err
-	}
-	maySuggestUpgrade(cmd)
-	i := cmdlog.NewSchemaInspect(ctx, client, s)
-	i.URL = flags.url
-	return format.Execute(cmd.OutOrStdout(), i)
-}
-
 // schemaFmtCmd represents the 'atlas schema fmt' subcommand.
 func schemaFmtCmd() *cobra.Command {
 	cmd := &cobra.Command{
