+++ sql/sqlite/inspect.go
@@ -21,100 +21,6 @@ type inspect struct{ *conn }
 
 var _ schema.Inspector = (*inspect)(nil)
 
-// InspectRealm returns schema descriptions of all resources in the given realm.
-func (i *inspect) InspectRealm(ctx context.Context, opts *schema.InspectRealmOption) (*schema.Realm, error) {
-	schemas, err := i.databases(ctx, opts)
-	if err != nil {
-		return nil, err
-	}
-	if len(schemas) > 1 {
-		return nil, fmt.Errorf("sqlite: multiple database files are not supported by the driver. got: %d", len(schemas))
-	}
-	if opts == nil {
-		opts = &schema.InspectRealmOption{}
-	}
-	var (
-		r    = schema.NewRealm(schemas...)
-		mode = sqlx.ModeInspectRealm(opts)
-	)
-	if mode.Is(schema.InspectTables) {
-		for _, s := range schemas {
-			tables, err := i.tables(ctx, nil)
-			if err != nil {
-				return nil, err
-			}
-			s.AddTables(tables...)
-			for _, t := range tables {
-				if err := i.inspectTable(ctx, t); err != nil {
-					return nil, err
-				}
-			}
-		}
-		sqlx.LinkSchemaTables(r.Schemas)
-	}
-	if mode.Is(schema.InspectViews) {
-		if err := i.inspectViews(ctx, r, nil); err != nil {
-			return nil, err
-		}
-	}
-	if mode.Is(schema.InspectTriggers) {
-		if err := i.inspectTriggers(ctx, r, nil); err != nil {
-			return nil, err
-		}
-	}
-	return schema.ExcludeRealm(r, opts.Exclude)
-}
-
-// InspectSchema returns schema descriptions of the tables in the given schema.
-// If the schema name is empty, the "main" database is used.
-func (i *inspect) InspectSchema(ctx context.Context, name string, opts *schema.InspectOptions) (*schema.Schema, error) {
-	if name == "" {
-		name = mainFile
-	}
-	schemas, err := i.databases(ctx, &schema.InspectRealmOption{
-		Schemas: []string{name},
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(schemas) == 0 {
-		return nil, &schema.NotExistError{
-			Err: fmt.Errorf("sqlite: schema %q was not found", name),
-		}
-	}
-	if opts == nil {
-		opts = &schema.InspectOptions{}
-	}
-	var (
-		r    = schema.NewRealm(schemas...)
-		mode = sqlx.ModeInspectSchema(opts)
-	)
-	if mode.Is(schema.InspectTables) {
-		tables, err := i.tables(ctx, opts)
-		if err != nil {
-			return nil, err
-		}
-		r.Schemas[0].AddTables(tables...)
-		for _, t := range tables {
-			if err := i.inspectTable(ctx, t); err != nil {
-				return nil, err
-			}
-		}
-		sqlx.LinkSchemaTables(schemas)
-	}
-	if mode.Is(schema.InspectViews) {
-		if err := i.inspectViews(ctx, r, opts); err != nil {
-			return nil, err
-		}
-	}
-	if mode.Is(schema.InspectTriggers) {
-		if err := i.inspectTriggers(ctx, r, nil); err != nil {
-			return nil, err
-		}
-	}
-	return schema.ExcludeSchema(r.Schemas[0], opts.Exclude)
-}
-
 func (i *inspect) inspectTable(ctx context.Context, t *schema.Table) error {
 	if err := i.columns(ctx, t); err != nil {
 		return err
